consolidated modifier under same interface signature (2 different argument options)
